feat(deps): outdated versions based on branch series
Use the minor version of a library used in an application branch
to verify whether the library version is outdated. App dependency
data is now stored as a JSON file containing structured information
about the library series used in each branch. Outdated versions in
each branch are highlighted along with the latest version in that
series, making a dedicated latest version column unnecessary.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>